Optimizations (#66)
* WIP: Fix for changing things to weight rather than priority

* Changed optimization schema to be weight

* The minimal generated test case for 11 conditions

* Changed all language files/parser to have weight instead of priority

* If no optimizations are provided then The task weight is based on time

* rename generated test case because I want to keep this file to de-bug with

* Fixed GenerateTestCase by removing battery

* Fix issue with Graph manager original hueristic distance creation.

* Fix for addition of battery to initial state in TestCaseGenerator

* Added limit to frontier size to increase search time and speed. Added exponent to Heuristic to make the search not go on forever in trivial cases.

* Change to many_paths to make it runnable and solvable.

* initialHeuristic move to TaskWeight

* TestCaseGenerator fix along with changing stopTime to be a double

+++ src/main/java/ai/promethean/Planner/Algorithm.java
@@ -15,5 +15,5 @@ public interface Algorithm {
      *
      * @return the system state
      */
-    SystemState run(SystemState initState, GoalState goalState, TaskDictionary tasks, StaticOptimizations optimizations, int stopTime, boolean activateCLF);
+    SystemState run(SystemState initState, GoalState goalState, TaskDictionary tasks, StaticOptimizations optimizations, double stopTime, boolean activateCLF);
 }
